+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-export const NotFound = () => {
+const NotFound = () => {
   return (
     <section aria-label="Content not found" className="pt-20 text-center">
       <h1 className="text-5xl">404 Not found</h1>
