+++ src/components/Blog/Intro.tsx
@@ -1,60 +1,51 @@
 import { Keyword } from "design-system/Keyword";
-import { Text } from "design-system/Text";
 
 export const BlogIntro = () => {
   return (
     <>
       <header>
-        <Text as="h1" $fontSize="1.5rem" my={4}>
-          Blog
-        </Text>
+        <h1 className="my-4 text-2xl">Blog</h1>
       </header>
 
       <div className="max-w-prose">
-        <Text mb={2} $fontWeight={300}>
+        <p className="font-light mb-2">
           Here I publish my solutions to coding challenges, and things I learn
           on my day to day job.
-        </Text>
+        </p>
 
-        <Text mb={2} $fontWeight={300}>
+        <p className="font-light mb-2">
           I also publish on Medium. You can find me over there as{" "}
-          <Text
-            as="a"
+          <a
             href="https://medium.com/@icjoseph"
             target="_blank"
             rel="noopener noreferrer"
-            $textColor="--yellow"
-            $fontWeight={400}
+            className="text-pale-yellow font-normal"
           >
             <i>@icjoseph</i>
-          </Text>
+          </a>
           .
-        </Text>
+        </p>
 
-        <Text as="h2" pt={1} mt={3} mb={2} $fontSize="1.125rem">
-          Coding Challenges
-        </Text>
+        <h2 className="pt-1 mt-3 mb-2 text-lg">Coding Challenges</h2>
 
-        <Text mb={2} $fontWeight={300}>
+        <p className="font-light mb-2">
           I put time into coding competitons and coding challengens. I focus
           mostly on <Keyword>Hash Code</Keyword>, <Keyword>Code Jam</Keyword>,{" "}
           <Keyword>Kick Start</Keyword>, and <Keyword>Advent of Code</Keyword>.
-        </Text>
+        </p>
 
-        <Text mb={2} $fontWeight={300}>
+        <p className="font-light mb-2">
           When solving coding challenges, I use the Rust programming language.
-        </Text>
+        </p>
 
-        <Text as="h2" pt={1} mt={3} mb={2} $fontSize="1.125rem">
-          Lessons Learned
-        </Text>
+        <h2 className="pt-1 mt-3 mb-2 text-lg">Lessons Learned</h2>
 
-        <Text mb={2} $fontWeight={300}>
+        <p className="font-light mb-2">
           In this blog, I take the opportunity to document, things I learn, or
           best practices to use when working with front end and back end
           frameworks, using <Keyword>JavaScript</Keyword>,{" "}
           <Keyword>TypeScript</Keyword>, and even <Keyword>Rust</Keyword>.
-        </Text>
+        </p>
       </div>
     </>
   );
