--- src/design-system/Keyword.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-import { theme } from "design-system/Global/theme";
-
-export const Keyword = styled.b`
-  display: inline;
-  font-weight: 400;
-  font-size: inherit;
-  color: ${theme.lightGreen};
-  color: var(--lightGreen);
-`;
