+++ src/design-system/Keyword.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from "react";
+
+export const Keyword = ({ children }: { children: ReactNode }) => (
+  <b className="inline font-normal text-pale-green">{children}</b>
+);
